+++ src/main/java/com/cafepos/decorator/Syrup.java
@@ -4,8 +4,10 @@ import com.cafepos.catalog.Priced;
 import com.cafepos.common.money;
 import com.cafepos.domain.product;
 
+import java.math.BigDecimal;
+
 public final class Syrup extends ProductDecorator {
-    private static final money SURCHARGE = money.of(0.40);
+    private static final money SURCHARGE = money.of(BigDecimal.valueOf(0.40));
     public Syrup(product base) { super(base); }
     @Override public String name() { return base.name() + " + Syrup"; }
     @Override public money price() {
