+++ src/main/java/com/cafepos/demo/week2Demo.java
@@ -8,11 +8,13 @@ import com.cafepos.domain.order;
 import com.cafepos.domain.orderIds;
 import com.cafepos.domain.simpleProduct;
 
+import java.math.BigDecimal;
+
 public final class week2Demo {
     public static void main(String[] args) {
         catalog catalog = new inMemoryCatalog();
-        catalog.add(new simpleProduct("P-ESP", "Espresso", money.of(2.50)));
-        catalog.add(new simpleProduct("P-CCK", "Chocolate Cookie", money.of(3.50)));
+        catalog.add(new simpleProduct("P-ESP", "Espresso", money.of(BigDecimal.valueOf(2.50))));
+        catalog.add(new simpleProduct("P-CCK", "Chocolate Cookie", money.of(BigDecimal.valueOf(3.50))));
         order order = new order(orderIds.next());
         order.addItem(new lineItem(catalog.findById("P-ESP").orElseThrow(), 2));
         order.addItem(new lineItem(catalog.findById("P-CCK").orElseThrow(), 1));
